# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a music subscription service server.",
        "title": "Music Subscription Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/users/{id}/subscription": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the subscription type of a user (ADMIN only). Types: FREE, PREMIUM.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update user subscription",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated subscription payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_subscriptions_handler.UpdateSubscriptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid payload or ID",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (ADMIN required)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/playlists/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information of a user's playlist by ID. Users can only get their own playlists.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playlists"
                ],
                "summary": "Get a playlist by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Playlist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_features_playlists_model.Playlist"
                        }
                    },
                    "400": {
                        "description": "Invalid playlist ID",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (Not owner of the playlist)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Playlist not found",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the title and description of a user's playlist. Users can only update their own playlists.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "playlists"
                ],
                "summary": "Update a playlist",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Playlist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated playlist payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_playlists_handler.CreatePlaylistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_features_playlists_model.Playlist"
                        }
                    },
                    "400": {
                        "description": "Invalid payload or ID",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (Not owner of the playlist)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Playlist not found",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a user's playlist by ID. Users can only delete their own playlists.",
                "tags": [
                    "playlists"
                ],
                "summary": "Delete a playlist",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Playlist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content (Playlist successfully deleted)"
                    },
                    "400": {
                        "description": "Invalid playlist ID",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (Not owner of the playlist)",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Playlist not found",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/tracks/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search active tracks by title or artist name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Search tracks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/music-service_internal_features_tracks_model.Track"
                            }
                        }
                    },
                    "400": {
                        "description": "Query parameter is required",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/music-service_internal_core_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_features_playlists_handler.CreatePlaylistRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
        "internal_features_subscriptions_handler.UpdateSubscriptionRequest": {
            "type": "object",
            "required": [
                "subscription_type"
            ],
            "properties": {
                "subscription_type": {
                    "type": "string",
                    "enum": [
                        "FREE",
                        "PREMIUM"
                    ]
                }
            }
        },
...
        "music-service_internal_core_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "music-service_internal_features_playlists_model.Playlist": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "music-service_internal_features_tracks_model.Track": {
            "type": "object",
            "properties": {
                "album_id": {
                    "type": "integer"
                },
                "album_title": {
                    "type": "string"
                },
                "artist_id": {
                    "type": "integer"
                },
                "artist_name": {
                    "description": "Join fields",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration_seconds": {
                    "type": "integer"
                },
                "file_url": {
                    "type": "string"
                },
                "genre_id": {
                    "type": "integer"
                },
                "genre_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "ВАЖНО: вводите с префиксом \"Bearer \". Пример: Bearer eyJhbGci...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}